update menu.cs to the correct format, it was used with prior adjustments that are no longer needed
I tried to use another methods to let the user press escape key whenever inside a user selection menu with inputs, this was a pain because I didn't know what was happening everytime the user pressed the escape key. Even tried to hardcode it, and print debug info, but this was all no longer needed.