removed unused code from a prior method I tried for trying to add esc key functionality to the input sections
I was very struggling with this class, I tried to add a input checker but the stupidity of me was not very useful. The Show method uses a list to use the arrow keys, so i tried to add escape key functionality and by doing this, it would break the actual code a bit. Instead of returning to the previous menu, I would see the first letter, be removed from a Menu name that you defined in the Show method.

This was a total pain, but worth the investment down the line, because I tried other methods and now have a workable code that is modular and can be used across the program. Only difference is, it is apart from the arrow key functionality, the esc key on the keyboard can be pressed to return to a previous menu, like I said in another commit summary lol.