refactored this pain code to look at
I'm trying to fix the selection menu showing the controls in the footer of the terminal, I looked at the code and saw it was basically the same two methods, only difference is. As a user it works, as admin it doesn't and I don't know why. Asked @Davidek523 for help